+++ src/encoded/static/components/graph.js
@@ -88,7 +88,6 @@ var Graph = module.exports.Graph = React.createClass({
     // jsonGraph: JsonGraph object containing nodes and edges.
     // graph: Initialized empty Dagre-D3 graph.
     convertGraph: function(jsonGraph, graph) {
-
         // graph: dagre graph object
         // parent: JsonGraph node to insert nodes into
         function convertGraphInner(graph, parent) {
@@ -278,5 +277,256 @@ var Graph = module.exports.Graph = React.createClass({
                 {this.props.children}
             </div>
         );
+    }
+});
+
+
+// This component should really be in experiments.js, but it's here temporarily
+// so that both experiment.js and dataset.js can use it without a cyclical dependency.
+
+var ExperimentGraph = module.exports.ExperimentGraph = React.createClass({
+    // Create nodes based on all files in this experiment
+    assembleGraph: function() {
+        var context = this.props.context;
+        var files = this.props.files || context.files;
+        var jsonGraph;
+
+        // Save list of files so we can click-test them later
+        this.fileList = files;
+
+        // Only produce a graph if there's at least one file with an analysis step
+        // and the file has derived from other files.
+        if (files && files.some(function(file) {
+            return file.derived_from && file.derived_from.length && file.step;
+        })) {
+            // Create an empty graph architecture
+            jsonGraph = new JsonGraph('');
+
+            // Create nodes for the replicates
+            context.replicates.forEach(function(replicate) {
+                jsonGraph.addNode(replicate.biological_replicate_number, replicate.biological_replicate_number, 'pipeline-replicate', 'rp');
+            });
+
+            // Add files and their steps as nodes to the graph
+            files.forEach(function(file) {
+                var fileId = file['@id'];
+                var replicateNode = file.replicate ? jsonGraph.getNode(file.replicate.biological_replicate_number) : null;
+
+                // Assemble a single file node; can have file and step nodes in this graph, so use 'fi' type
+                // to show that this is a file node.
+                jsonGraph.addNode(fileId, file.accession + ' (' + file.output_type + ')',
+                    'pipeline-node-file' + (this.state.infoNodeId === fileId ? ' active' : ''), 'fi', replicateNode);
+
+                // If the node has parents, build the edges to the analysis step between this node and its parents
+                if (file.derived_from && file.derived_from.length && file.step) {
+                    var step = file.step.analysis_step;
+                    var stepId = step['@id'] + '&' + file['@id'];
+
+                    // Insert a node for the analysis step, with an ID combining the IDs of this step and the file that
+                    // points to it; there may be more than one copy of this step on the graph if more than one
+                    // file points to it, so we have to uniquely ID each analysis step copy with the file's ID.
+                    // 'as' type identifies these as analysis step nodes. Also add an edge from the file to the
+                    // analysis step.
+                    jsonGraph.addNode(stepId, step.analysis_step_types.join(', '),
+                        'pipeline-node-analysis-step' + (this.state.infoNodeId === stepId ? ' active' : ''), 'as', replicateNode);
+                    jsonGraph.addEdge(stepId, fileId);
+
+                    // Draw an edge from the analysis step to each of the derived_from files
+                    file.derived_from.forEach(function(derived) {
+                        jsonGraph.addEdge(derived['@id'], stepId);
+                    });
+                }
+            }, this);
+        }
+        return jsonGraph;
+    },
+
+    getInitialState: function() {
+        return {
+            infoNodeId: '' // @id of node whose info panel is open
+        };
+    },
+
+    detailNodes: function(jsonGraph, infoNodeId) {
+        var meta;
+        var selectedFile;
+
+        // Find data matching selected node, if any
+        if (infoNodeId) {
+            var node = jsonGraph.getNode(infoNodeId);
+            if (node) {
+                switch(node.type) {
+                    case 'fi':
+                        // The node is for a file
+                        selectedFile = _(this.fileList).find(function(file) {
+                            return file['@id'] === infoNodeId;
+                        });
+
+                        if (selectedFile) {
+                            var analysisStep = selectedFile.step && selectedFile.step.analysis_step;
+                            meta = (
+                                <dl className="key-value">
+                                    {selectedFile.file_format ?
+                                        <div data-test="format">
+                                            <dt>Format</dt>
+                                            <dd>{selectedFile.file_format}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedFile.output_type ?
+                                        <div data-test="output">
+                                            <dt>Output</dt>
+                                            <dd>{selectedFile.output_type}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedFile.paired_end ?
+                                        <div data-test="pairedend">
+                                            <dt>Paired end</dt>
+                                            <dd>{selectedFile.paired_end}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedFile.replicate ?
+                                        <div data-test="replicate">
+                                            <dt>Associated replicates</dt>
+                                            <dd>{'(' + selectedFile.replicate.biological_replicate_number + ', ' + selectedFile.replicate.technical_replicate_number + ')'}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedFile.submitted_by && selectedFile.submitted_by.title ?
+                                        <div data-test="submitted">
+                                            <dt>Added by</dt>
+                                            <dd>{selectedFile.submitted_by.title}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedFile.date_created ?
+                                        <div data-test="datecreated">
+                                            <dt>Date added</dt>
+                                            <dd>{selectedFile.date_created}</dd>
+                                        </div>
+                                    : null}
+
+                                    {analysisStep && analysisStep.software_versions && analysisStep.software_versions.length ?
+                                        <div>
+                                            <dt>Software</dt>
+                                            <dd>
+                                                {<ul>
+                                                    {analysisStep.software_versions.map(function(version, i) {
+                                                        return (
+                                                            <span>
+                                                                {i > 0 ? ', ' : ''}
+                                                                <a href={version.software['@id']}>{version.software.name}</a>
+                                                            </span>
+                                                        );
+                                                    })}
+                                                </ul>}
+                                            </dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedFile.pipeline ?
+                                        <div data-test="pipeline">
+                                            <dt>Pipeline</dt>
+                                            <dd><a href={selectedFile.pipeline['@id']}>{selectedFile.pipeline.title}</a></dd>
+                                        </div>
+                                    : null}
+                                </dl>
+                            );
+                        }
+
+                        break;
+
+                    case 'as':
+                        // The node is for an analysis step
+                        var analysisStepId = node.id.slice(0, node.id.indexOf('&'));
+                        selectedFile = _(this.fileList).find(function(file) {
+                            return file.step && file.step.analysis_step['@id'] === analysisStepId;
+                        });
+
+                        if (selectedFile) {
+                            var selectedStep = selectedFile.step.analysis_step;
+
+                            meta = (
+                                <dl className="key-value">
+                                    {selectedStep.input_file_types && selectedStep.input_file_types.length ?
+                                        <div data-test="inputtypes">
+                                            <dt>Input file types</dt>
+                                            <dd>{selectedStep.input_file_types.join(', ')}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedStep.output_file_types && selectedStep.output_file_types.length ?
+                                        <div data-test="outputtypes">
+                                            <dt>Output file types</dt>
+                                            <dd>{selectedStep.output_file_types.join(', ')}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedStep.qa_stats_generated && selectedStep.qa_stats_generated.length ?
+                                        <div data-test="steptypes">
+                                            <dt>QA statistics</dt>
+                                            <dd>{selectedStep.qa_stats_generated.join(', ')}</dd>
+                                        </div>
+                                    : null}
+
+                                    {selectedStep.software_versions && selectedStep.software_versions.length ?
+                                        <div>
+                                            <dt>Software</dt>
+                                            <dd>
+                                                {<ul>
+                                                    {selectedStep.software_versions.map(function(version, i) {
+                                                        return (
+                                                            <span>
+                                                                {i > 0 ? ', ' : ''}
+                                                                <a href={version.software['@id']}>{version.software.name}</a>
+                                                            </span>
+                                                        );
+                                                    })}
+                                                </ul>}
+                                            </dd>
+                                        </div>
+                                    : null}
+                                </dl>
+                            );
+                        }
+
+                        break;
+
+                    default:
+                        break;
+                }
+            }
+        }
+
+        return meta;
+    },
+
+    handleNodeClick: function(e, nodeId) {
+        e.stopPropagation(); e.preventDefault();
+        this.setState({infoNodeId: this.state.infoNodeId !== nodeId ? nodeId : ''});
+    },
+
+    render: function() {
+        var context = this.props.context;
+
+        // Build node graph of the files and analysis steps with this experiment
+        var jsonGraph = this.assembleGraph();
+        if (jsonGraph) {
+            var meta = this.detailNodes(jsonGraph, this.state.infoNodeId);
+            return (
+                <div>
+                    <h3>Pipeline for {context.accession}</h3>
+                    <Graph graph={jsonGraph} nodeClickHandler={this.handleNodeClick}>
+                        <div className="graph-node-info">
+                            {meta ? <div className="panel-insert">{meta}</div> : null}
+                        </div>
+                    </Graph>
+                </div>
+            );
+        } else {
+            return null;
+        }
     }
 });
